requirements.txt
asgiref==3.7.2
asttokens==2.4.1
devtools==0.12.2
Django==4.2.7
django-simple-history==3.4.0
djangorestframework==3.14.0
executing==2.0.1
Pillow==10.1.0
Pygments==2.17.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
